[NFCI][SYCL] Eliminate `sycl/builtins_utils_*.hpp`
Split the helper they used to provide between the builtins
implementation and `sycl/detail/type_traits/vec_marray_traits.hpp`.
