+++ sycl/include/sycl/ext/oneapi/experimental/builtins.hpp
@@ -10,7 +10,6 @@
 
 #include <sycl/aliases.hpp>                    // for half
 #include <sycl/builtins.hpp>                   // for to_vec2
-#include <sycl/builtins_utils_vec.hpp>         // for to_vec, to_marray...
 #include <sycl/detail/defines_elementary.hpp>  // for __SYCL_ALWAYS_INLINE
 #include <sycl/detail/generic_type_traits.hpp> // for is_svgenfloath, is_sv...
 #include <sycl/detail/memcpy.hpp>              // detail::memcpy
